feat[DE]: Create ConsoleUI class
